BLINK: Added notes to blink(1) lsusb v output
Linux Device Drivers 3rd Edition had some illuminating details regarding "Endpoints".  In an attempt to better understand blink(1)'s sole Endpoint, I appended inline C-style comments to some of the output lines from lsusb -v (as it pertained to the blink(1) device).

The interesting note is the realization that the max # of bytes the blink(1) can handle at once also happens to match the size of the data array the blink(1) API utilizes to send messages to the device.  

-Hark