BLINK: Implemented read_buff()
Current problems with blink(1) device not... well, blinking... appear to be from unacknowledged communication from the device.  If this doesn't work, we'll try to set transfer flags to 0x204 (like the functional packet capture).  If that doesn't work, we'll have to 'read'(?) the reports.  Maybe we're not speaking the *exact* same language as blink(1).

WRITTEN OFFLINE.  TEST LATER!

-Hark