key: BIG LKM.c update
The following major changes were made to myKeyLoggingModule.c:
	- Extricated macros into external header file at the root of the repo
	- Added default MACROs to aid in File I/O
	- Added struct members to aid in File I/O
	- Defined open_log()
	- Defined close_log()
	- Modified init() to call open_log()
	- Modified exit() to call close_log()

Everything appears to be functional.  File is created with the proper permissions with the proper absolute filename.  DEBUG LKM logging indicates that everything is functional.

Next step, implement write() to enable *actual* logging.

NOTE TO THE FUTURE: Propogate that into the other LKMs within Peke_Ito
Changes to be committed:
	modified:   myKeyLoggingModule.c