key: Fixed another edge case BUG
This particular BUG had to do with the user attempting to read from the device when it's empty.  I never check the length of the device's buffer prior to a read.  Now, I do.  copy_to_user() won't execute if the device buffer is empty.

This fixes the follwing error (for situations in which the device buffer length is 0):
	<<<ERROR>>> device_read - copy_to_user() failed to copy all the bytes

Changes to be committed:
	modified:   myKeyLoggingModule.c