process: Enchanced LKM
Added facility within the LKM to determine what the max process name
was.  This is important for finding process names.  I added a default
macro based on my observations (confirmed by the LKM itself).  I also
added a least-length conditional to name_lookup() to verify that
name_lookup(git-credential-cache) wouldn't fail because it's listed as
"git-credential-".

Changes to be committed:
	modified:   myProcessModule.c